+++ db/migrate/20160219162059_add_grid_variables_count_to_exports.rb
@@ -1,0 +1,5 @@
+class AddGridVariablesCountToExports < ActiveRecord::Migration
+  def change
+    add_column :exports, :grid_variables_count, :integer, null: false, default: 0
+  end
+end